pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "factory-boy>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]

[tool.ruff]
line-length = 105
exclude = [
    "tests",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]